[test] Fix Android crash by moving TracedReference to stack
This test was failing on the Pixel 7. On Android we run unittests
with the system memory allocator. It seems that the allocator on
Android uses the top-byte for storing additional bytes. However,
in our test we store the pointer in the CppHeapPointerTable, where we
DCHECK that the highest 16-bits are all 0.

This CL avoids this by storing v8::TracedReference on the stack
instead of the heap. That way we ensure that even on Android the
highest bits of the address of v8::TracedReference are clear.

Bug: 375121012
Change-Id: Iba9a969cca87858c3d47da45db12caf95df14aac
Cq-Include-Trybots: luci.v8.try:v8_android_arm64_p7_rel