[gc] Use exponential decay for allocation speed
This CL introduces SmoothedBytesAndDuration class to
compute exponentially smoothed throughput (based on time, not samples).
This is used to track allocation speed in GCTracer, with
a half life of 100ms.

Previously, we would use BoundedAverageSpeed (moving window with no
decay) with either
- kThroughputTimeFrame=5s
- full window, affected by buffer size = 10 samples
In a somewhat inconsistent way.
This CL gets rid of the distinction and imposes a single way to compute
the average on allocation throughput.

Bug: 42203776
Change-Id: I304cd13199d339440e8cea1ebba4003ddc49720a