[maglev] speeling: More precision in KNA compatibility check
When the loaded context slots at loop antry are empty we do not need to
track aliasing state of contexts at all. The aliasing state is only
relevant to decide if a newly written value can alias with an already
existing entry in the cache. Therefore the known node aspects with no
context slots cached is trivially compatible with any contet alias
analysis state.

This improves over the fixes interoduced in
https://crrev.com/c/5982989.

Bug: v8:7700
Change-Id: Ic04f92ba5cd55a90e221b659445d77d3843740cb