[turboshaft] Disable FORCE_UNROLL_LOOP
Because of a bug in LLVM, Clang sometimes fails to unroll the loop
(and complains about failing). So, we're disabling this for now, and
we'll check back in a few months to see if this has been fixed.

This CL might slightly slow down Turboshaft compilation and increase
Turboshaft binary size, but I don't think that any of this will be
impactful enough to be measured by our bots (I'd guess a 0.1% impact
max, although I haven't measured).

Bug: chromium:377144577
Change-Id: Ia7ac4e41e134770f16a3dd8616e8ba4c051b7c38