[wasm] Hardcode jump table instruction bytes on ia32/x64
The CL is also padding the jump table entries to 8 bytes on x64, since
IIUC, AMD doesn't guarantee that unaligned writes are atomic on the same
cache line, but have a stricter requirement that they're in a single
qword.

This will allow us to remove the dependency on the MacroAssembler.
This is useful for two reasons:
1) it gives more control over the instruction layout
2) we want to use the WritableJitAllocation class to perform the writes,
   which we'd have to pass into the MacroAssembler too.

Bug: 42203297
Change-Id: I9c34573bc3748dfbdc58ea6f29081f461abe7793