[sandbox][string] Add String::Utf8Length and String::EncodeUtf8
Currently, we have a number of different implementations of Utf8
encoding in V8: in v8::String::WriteUtf8 (the public API),
v8::internal::String::ToCString (an internal API), in the Wasm runtime
(EncodeWtf8), and in some debugging code. Furthermore, many of these
implementations are not sandbox-compatible: they typically work by first
computing the required length of the output buffer, then encoding the
string into the allocated buffer without further bounds-checking. This
is no longer correct as the length or content of the string can change
in between. This CL chain aims to unify the Utf8 encoding code and make
it sandbox-compatible at the same time.

This CL is a first step towards that goal. It introduces new helper
functions String::Utf8Length and String::EncodeUtf8 into the internal
String objects and uses them for String::ToCString. Follow-up CLs will
migrate the other users of Utf8 encoding to these APIs.

The current implementation of String::EncodeUtf8 is relatively simple.
Further optimizations will be done as necessary when migrating the other
(potentially more performance-sensitive) users of Utf8 encoding.

Bug: 373485796, 376071292
Change-Id: I798b288c654cb35b80b069916a781811a6cae09b