[leaptiering] Leap-tier-up
Implement tiering up solely using the dispatch table instead of
checking the tiering state on function calls or entries.

To do so, we remove the TieringState (from the FBV) completely and
instead rely on two pieces:

* To indicate if a tiering is in progress we use a flag called
  `tiering_in_progress` on the FBV.
* To request a tiering, i.e., to start a compilation job, we install a
  tiering trampoline into the dispatch table.

Tiering trampolines are installed by overriding the `entrypoint`
field in the dispatch table, but leaving the `code` part untouched.

Logging function executions is also handled using the same mechanism.

Bug: 42204201

Change-Id: If99105ac3c1c0f55447bfd928f0f4c03c4204eec