[sandbox][string] Harden String::ToCString against in-sandbox corruption
The String::ToCString method is one of a few implementations of Utf8
encoding currently available in V8. Like many of them, it is not
currently sandbox-compatible: it works by first computing the required
length of the output buffer, then encoding the string into the allocated
buffer without further bounds-checking. This is not correct as the
length or content of the string can change in between. This CL hardens
String::ToCString to correctly handle this case through an SBXCHECK.

Follow-up CLs will refactor the other places that perform Utf8 encoding
to be robust against in-sandbox memory corruption and ideally be able to
share more code instead of re-implementing Utf8 encoding.

Bug: 376071292
Change-Id: I798b288c654cb35b80b069916a781811a6cae09b