[sandbox] Harden JsonStringifier::SerializeString
The previous logic for serializing a String object to JSON was
implicitly assuming that the length of a String could not change (as it
was loading the length twice, once explicitly, once when obtaining the
FlatContent of the string). This assumption is no longer correct with
the sandbox, so with this CL, we're removing the first load of the
length and only using the length of the string's CharVector.

Bug: 376071292
Change-Id: I2dd43d5df79e269356070763976ddad89ce49659