[objects] Remove GetHeap() from JSObjectShortPrint()
JSObjectShortPrint() had special detection for broken references. This
CL removes this becomes it doesn't provide much value anymore.

The GetHeap() call was using GetHeapFromWritableObject() under
the hood. This was triggering DCHECKs for shared objects and
because of that all JS objects in the shared heap needed special
handling in JSObjectShortPrint().

Bug: 376564609
Change-Id: If0a5cb56e70cac19bfd71c9f4ab027075731d19d