[wasm][fastapi] Handle return value of backup regular API call
Fast API calls call directly from generated code to C++. When a
parameter is not within range of the type expected by the C++ target,
then a regular API call is used, which is expected to throw an
exception. However, this expectation depends on user code and may not
hold, as in a regular API call, parameters get converted from
WebAssembly values to JavaScript values in the wasm-to-js wrapper, and
after that conversion parameter may actually be within the range of the
expected C++ type.

With this CL the return value of the regular API call gets handled
correctly if the API call actually does not throw.

R=clemensb@chromium.org

Bug: 375270509
Change-Id: Ibbecfabaf34879a989fb5f90de2b9616f631c879