[ICs] Fix faux polymorphism in ARES-6 / ML
A KeyedStoreIC where we store to Bcolj in mmul was polymorphic (with
HOLEY_DOUBLE_ELEMENTS & HOLEY_SMI_ELEMENTS) even though it strictly
woudln't need to be.

What caused the polymorphism was:
1) we create the first Bcolj with HOLEY_SMI_ELEMENTS. there's no AllocationSite at this point.
2) KeyedStoreIC, we have no feedback
3) Bcolj turns to HOLEY_DOUBLE_ELEMENTS
4) KeyedStoreIC some more, at some point we have feedback; the IC is
   now monomorphic with HOLEY_DOUBLE_ELEMENTS
5) create next Bcolj, now we have an AllocationSite but it doesn't
   know what happened before, so it has HOLEY_SMI_ELEMENTS
6) IC becomes polymorphic with HOLEY_DOUBLE_ELEMENTS &
   HOLEY_SMI_ELEMENTS
7) the AllocationSite gets updated to HOLEY_DOUBLE_ELEMENTS
8) all created Bcoljs have HOLEY_DOUBLE_ELEMENTS, but the IC is still
   polymorphic even though it doesn't need to be any more.

The solution is: When the IC is monomorphic with
HOLEY_DOUBLE_ELEMENTS and the incoming object has HOLEY_SMI_ELEMENTS,
check whether there's an AllocationSite which could be transitioned
to HOLEY_DOUBLE_ELEMENTS. If that works, keep the IC monomorphic.


Change-Id: I01c4b35469009bbe743a70b7a92d8e02f1e8bf7d