[heap, builtins] Check bucket bounds for remembered set
SlotSet used to only store the number of allocated buckets in debug
builds. This CL enables the counter also on release builds and also
uses the counter in bottlenecks that compute bucket indicies. This
solves a heap sandbox issue with OOB access on corrupted slot offsets.

The CL reuses the existing debug counter for the CHECK. This already
turns out to be neutral on competitive benchmarks.

We can improve this by moving the counter next to the actual bucket
array pointer. This would make SlotSet roughly similar to a vector
(size + dynamic backing) and make sure that size is always in L1 this
way. The refactoring is left for future work as this requires adopting
how SlotSet works on background threads which currently only uses CAS
on the SlotSet pointer (which at this point includes the bucket
counter).

Bug: 329345899
Change-Id: Iceb522f3313d7e9f57f511d1b60c8983502ed955